requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
finmind==1.8.2
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
loguru==0.7.3
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
multidict==6.6.4
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyecharts==2.0.8
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
seaborn==0.13.2
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.48.1
ta==0.5.25
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
win32_setctime==1.2.0
yarl==1.20.1
zipp==3.23.0
